added water params object
