+++ src/index.tsx
@@ -9,12 +9,8 @@
  * It can be left blank if not needed.
  */
 
-import GfiTsWaterParams from './direflow-components/gfi-ts-water-params';
+import GfiTsWaterParams from "./direflow-components/gfi-ts-water-params";
 
 GfiTsWaterParams.then((element) => {
-
-  /**
-   * Access DOM node when it's mounted
-   */
-  console.log('gfi-ts-water-params is mounted on the DOM', element);
+	console.log("gfi-ts-water-params is mounted on the DOM", element);
 });
